refactor: remove old autocompleter
