+++ src/app/app.ts
@@ -86,7 +86,7 @@ import { CloudPath } from "./util/cloudPath"
 import { authenticate, authenticationCommandGroup } from "./auth"
 import { formatTable } from "../framework/util"
 import dedent from "dedent"
-import { fsCommands } from "./featureInterfaces/fs/fs"
+import { fsCommands } from "./featureInterfaces/fs"
 import { syncCommand } from "./featureInterfaces/syncInterface"
 import { driveMountingCommand } from "./featureInterfaces/driveMountingInterface"
 import { webdavCommandGroup } from "./featureInterfaces/webdavInterface"
